pyproject.toml
[project]
name = "cuzn"
version = "0.0.1"
authors = [
  { name="Scott Houseman", email="scott@houseman.co.za" },
]
description = "A Python client for the Braze Customer Engagement Platform REST API"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
  "requests",
  "python-dotenv",
  "ujson"
]

[project.optional-dependencies]

dev = [
  "black",
  "flake8",
  "isort",
  "mypy",
  "pyupgrade",
  "pre-commit",

  # Mypy type stubs
  "types-ujson",
  "types-requests",

  # Test dependencies
  "pytest",
  "pytest-cov[all]",
  "pytest-mock",

  # Conflict resolutions
  'virtualenv < 20.16.3; python_version < "3.8"' # Resolve conflict: importlib-metadata<4.3,>=1.1.0 (from flake8==5.0.4->cuzn (pyproject.toml))
]

[project.urls]
repository = "https://github.com/houseman/cu-zn"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["cuzn*"]
namespaces = false

[tool.isort]
profile = "black"
src_paths = [
    "src",
    "test"
]
honor_noqa = true

[tool.mypy]
files = "src"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
testpaths = [
    "tests"
]
addopts = "--cov=src/cuzn --cov-report html --cov-report term"
